Improve execution speed with delayed imports
The python yaml modules takes about 2.5 seconds to load on
fuegotest.org.  Reduce processing time for most requests, by
only loading json or yaml when they are actually used.
In the case of yaml, it is not used ever, at the moment.

Add a NOTES files about cgi execution speed

Signed-off-by: Tim Bird <tbird20d@yahoo.com>