Return request_id from put_request
Have put_request return the request_id, so that the caller
can have a token to use to query the request status.
Previously we were returning the filename, but make the
output more structured so that it can be processed automatically.

While we're at it, change the request_id from a filename to
a string (without the '.json' extension).  The fserver clients
shouldn't be exposed to fserver internals, where possible.

Signed-off-by: Tim Bird <tbird20d@yahoo.com>