Add support for querying boards
Add do_query_boards, to allow support for 'ftc list-boards -r'
This supports obtaining a list of boards using wildcards
in criteria for host and board attributes of the board definition.

For now, use only 'host' and 'board' as attributes, but that
may expand in the future to allow for more nuanced requests.

Signed-off-by: Tim Bird <tbird20d@yahoo.com>