Report error when run data cannot be read
I was manually deleting files, and ended up creating a dangling
symlink.   This caused a python exception in show_run_table.
Avoid the exception by adding a try/except clause, and showing
a proper error message in the table, if there are problems
reading the run json file.

Also, show timestamp for the run.

Signed-off-by: Tim Bird <tbird20d@yahoo.com>