Add debug logging to fserver.py
If the debug variable is set, do extra debug output to the log
and use a special instrumented cgi.py module (named mycgi,
located in the source directory).  If you are not running fserver.py
on fuegotest.org, then you probably need to adjust the python
path to reflect the installation path of the code.

Signed-off-by: Tim Bird <tbird20d@yahoo.com>