Eliminate exception in html run table, due to test_name
Older versions of function.sh in the Fuego core were using
the field name of 'test_name' in the pre-run run.json file,
to hold the test name.  The correct field name is "name"
Due to this, some run.json files has the test name in "test_name"
and some in "name".

Make the parsing of the run.json data more robust, and handle
this broken case.

Signed-off-by: Tim Bird <tbird20d@yahoo.com>