utils: add some utils for managing the live script
Add some simple shell scripts to check the status of the live cgi
script for fserver.py, versus the file in the development directory.

Use the 'diff..' script to see if the live script needs to be updated.
Use the 'push..' script to update the live script from the development
version.

Signed-off-by: Tim Bird <tim.bird@sony.com>